Update getFullfullPoolAdvertisementTitle in poolUtils with abbr wrapper function
